# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /rpc/debug/options:
    get:
      operationId: RpcController_debugOptions
      parameters:
        - name: project
          required: false
          in: query
          description: 用于加载历史 action 和在线设备的功能组
          schema:
            type: string
      responses:
        '200':
          description: 手动 RPC 调试可选功能组、历史 action 与在线设备成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RpcDebugOptions'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 手动 RPC 调试可选功能组、历史 action 与在线设备
      tags:
        - rpc
      description: 手动 RPC 调试可选功能组、历史 action 与在线设备。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
  /rpc/debug/invoke/{project}/{action}:
    post:
      operationId: RpcController_debugInvoke
      parameters:
        - name: project
          required: true
          in: path
          schema:
            type: string
        - name: action
          required: true
          in: path
          schema:
            type: string
        - name: clientId
          required: false
          in: query
          description: 可选目标设备 ID；省略时在功能组在线设备中轮询。
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvokeDto'
      responses:
        '200':
          description: 使用后台权限手动发起 RPC 调试调用成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvokeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 使用后台权限手动发起 RPC 调试调用
      tags:
        - rpc
      description: 使用后台权限手动发起 RPC 调试调用。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
  /rpc/invoke/{project}/{action}:
    post:
      operationId: RpcController_invoke
      parameters:
        - name: project
          required: true
          in: path
          schema:
            type: string
        - name: action
          required: true
          in: path
          schema:
            type: string
        - name: clientId
          required: false
          in: query
          description: 可选目标设备 ID；省略时在功能组在线设备中轮询。
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvokeDto'
      responses:
        '201':
          description: 调用 project 内在线设备执行 action(可选 ?clientId 指定设备)成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvokeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - accessToken: []
      summary: 调用 project 内在线设备执行 action(可选 ?clientId 指定设备)
      tags:
        - rpc
      description: '调用 project 内在线设备执行 action(可选 ?clientId 指定设备)。使用 Authorization: Bearer rk_… 访问令牌，并校验功能组作用域。'
  /rpc/clientQueue:
    get:
      operationId: RpcController_clientQueue
      parameters:
        - name: project
          required: true
          in: query
          schema:
            type: string
        - name: clientId
          required: false
          in: query
          description: 可选设备 ID；提供时只查询该设备在线状态。
          schema:
            type: string
      responses:
        '200':
          description: project 内在线设备(或指定 clientId 的在线状态)成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientQueueResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - accessToken: []
      summary: project 内在线设备(或指定 clientId 的在线状态)
      tags:
        - rpc
      description: 'project 内在线设备(或指定 clientId 的在线状态)。使用 Authorization: Bearer rk_… 访问令牌，并校验功能组作用域。'
...
info:
  title: R2RPC API
  description: R2RPC 管理控制面、调用方 RPC 与监控查询的 HTTP API。设备长连接协议使用 WebSocket，另见项目协议文档。
  version: 0.1.0
  contact:
    name: R2RPC Contributors
    url: https://github.com/rezoch340/R2RPC
    email: ''
  license:
    name: UNLICENSED
    url: https://github.com/rezoch340/R2RPC/blob/main/LICENSE
...
servers:
  - url: /
    description: 当前部署
components:
  securitySchemes:
    adminJwt:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: 后台登录接口签发的管理员 JWT。
    accessToken:
      scheme: bearer
      bearerFormat: R2RPC Access Token
      type: http
      description: 访问令牌管理页面生成的 rk_ 前缀 Access Token。
  schemas:
...
    InvokeDto:
      type: object
      properties:
        timeoutSeconds:
          type: number
          description: 超时秒数
          default: 20
        payload:
          type: object
          description: 调用负载(扁平对象)
      required:
        - payload
    ErrorResponse:
      type: object
      required:
        - statusCode
        - message
        - timestamp
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          oneOf:
            - type: string
              example: 请求参数非法
            - type: array
              items:
                type: string
              example:
                - name must be a string
            - type: object
              additionalProperties: true
              description: Nest HttpException 的结构化响应，例如 statusCode/message/error。
        timestamp:
          type: string
          format: date-time
...
    Project:
      type: object
      required:
        - id
        - name
        - enabled
        - createdAt
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: cn-nodes
        description:
          type: string
          nullable: true
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
...
    RpcDebugOptions:
      type: object
      required:
        - projects
        - actions
        - clientIds
      properties:
        projects:
          type: array
          items:
            $ref: '#/components/schemas/Project'
        actions:
          type: array
          items:
            type: string
        clientIds:
          type: array
          items:
            type: string
    InvokeResponse:
      type: object
      required:
        - requestId
        - clientId
        - is_ok
        - status
        - httpCode
        - latencyMs
      properties:
        requestId:
          type: string
          format: uuid
        clientId:
          type: string
          nullable: true
        is_ok:
          type: boolean
        status:
          type: string
        httpCode:
          type: integer
          description: 设备业务结果码；不等同于本次 HTTP 传输状态码。
        latencyMs:
          type: integer
          minimum: 0
        payload:
          description: 任意可 JSON 序列化的值
          nullable: true
        error:
          type: string
          nullable: true
    ClientQueueResponse:
      oneOf:
        - type: object
          required:
            - clientId
            - online
          properties:
            clientId:
              type: string
            online:
              type: boolean
        - type: object
          required:
            - project
            - online
          properties:
            project:
              type: string
            online:
              type: array
              items:
                type: string
...
  responses:
    BadRequest:
      description: 请求参数、路径参数或查询参数不符合契约。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 凭证缺失、无效、过期、撤销或对应账号已不存在。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 凭证有效，但账号、功能组或 RBAC 权限不允许本次操作。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: 服务端发生未预期错误。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'